update control.spec.js integration tests. async mocks and spies close fate enables openbutton and lock button
