+++ src/controls/Controls.js
@@ -16,3 +16,4 @@ const Controls = props => {
 };
 
 export default Controls;
+
