+++ src/dashboard/Dashboard.spec.js
@@ -18,3 +18,4 @@ describe('Dashboard', () => {
         expect(getByText(/locked/i));
     })
 })
+
